Pagination logici geliştirildi ve  bazı varlıklar özelinde hem repo h…

+++ Core/Application/Features/CQRS/Results/CarResults/GetAllCarsQueryResult.cs
@@ -8,8 +8,8 @@
         public string CoverImageUrl { get; set; }
         public int Km { get; set; }
         public string Transmission { get; set; }
-        public byte Luggage { get; set; }
-        public byte Seat { get; set; }
+        public int Luggage { get; set; }
+        public int Seat { get; set; }
         public string Fuel { get; set; }
         public string BigImageUrl { get; set; }
     }
